geo-service/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
certifi==2020.4.5.1       # via requests
chardet==3.0.4            # via requests
click==7.1.1              # via flask
flask-cors==3.0.8         # via -r requirements.in
flask==1.1.1              # via -r requirements.in, flask-cors
googlemaps==4.4.1         # via -r requirements.in
gunicorn==20.0.4          # via -r requirements.in
idna==2.9                 # via requests
itsdangerous==1.1.0       # via flask
jinja2==2.11.1            # via flask
markupsafe==1.1.1         # via jinja2
python-dotenv==0.13.0     # via -r requirements.in
requests==2.23.0          # via googlemaps
six==1.14.0               # via flask-cors
urllib3==1.25.9           # via requests
werkzeug==1.0.0           # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
